implemented mem_write_32 function

+++ src/vm/memory/mem_access.h
@@ -8,7 +8,7 @@ uint16_t mem_fech_16(dtptr_t addr);
 uint32_t mem_fech_32(dtptr_t addr);
 
 //write memory functions
-void mem_write_16(dtptr_t addr, uint8_t data);
-void mem_write_32(dtptr_t addr, uint8_t data);
+void mem_write_16(dtptr_t addr, uint16_t data);
+void mem_write_32(dtptr_t addr, uint32_t data);
 
 #endif //! __MEMORY_ACCESS_H__
